// __generated__/getPurchaseHistory.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getPurchaseHistory
// ====================================================

export interface getPurchaseHistory {
  getPurchaseHistory: number[];
}

export interface getPurchaseHistoryVariables {
  seriesId: number;
}